Добавлены методы удаления типов продуктов и обновлены контроллеры и сервисы для работы с типами продуктов. Реализована логика удаления с учетом S3-хранилища для изображений. Обновлены SQL-запросы для корректного отображения данных.

+++ src/product-main/dto/create-product-main.dto.ts
@@ -110,7 +110,7 @@ export class CreateProductMainDto {
 
   @ApiProperty({ description: 'Калории', example: 270 })
   @Transform(({ value }) => {
-    const num = parseInt(value);
+    const num = parseFloat(value);
     return isNaN(num) ? 0 : num;
   })
   @IsNumber()
